[#3731] Extract shared MembershipResolvers; clarify SetRole save is updated_at-only
PR review follow-ups on the memberships adapters (#3731):

- De-dup: resolve_org/resolve_customer were byte-identical across the three
  colonel adapters (Add/Remove/SetMembershipRole). Extract into a shared
  MembershipResolvers concern. Kept in the membership namespace (not the
  global colonel Base) because the extid-first lookup order is
  membership-specific — transfer_domain.rb resolves orgs load-first.

- change_role! save: change_role! already persists the role field via
  materialize_for_role!, so Add correctly does not re-save. SetRole's save
  only stamps updated_at — comment now says so, resolving the apparent
  Add/SetRole inconsistency.

- normalize_email on an extid is a safe pass-through (extids are lowercase
  ASCII); documented in the resolver rather than guarded.

19 memberships op specs still green.
